# note-keeper-web/swagger.yaml
openapi: 3.0.3
info:
  title: NoteKeeper API
  description: |
    Open-source Evernote alternative with powerful features for notes and todos management.
    
    ## Features
    - **Notes Management**: Create, read, update, delete notes with Markdown & Mermaid support
    - **Todos Management**: Manage todos with descriptions, tags, priorities, and schedules
    - **Folder Structure**: Organize notes in folders and subfolders
    - **Tags & Priorities**: Categorize and prioritize your content
    - **Encryption**: Secure sensitive notes
    - **Reminders**: Set reminders for notes and todos
    - **Archive & Trash**: Archive old items and restore from trash
    - **Templates**: Create notes from predefined templates
    - **History**: Track note changes with version history
    - **Search**: Full-text search with saved queries
    - **Calendar**: Visualize todos with deadlines
    - **Analytics**: Productivity graphs and statistics
    - **Shortcuts**: Keyboard shortcuts for quick actions
    - **Integrations**: Send notifications to Telegram, DingTalk, and Email
  version: 3.0.0
  contact:
    name: NoteKeeper Support
    email: support@notekeeper.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.notekeeper.example.com/v1
    description: Production server
...
paths:
...
  /todos:
    get:
      tags:
        - Todos
      summary: Get all todos
      description: Retrieve a list of all todos with optional filtering
      operationId: getTodos
      parameters:
        - name: completed
          in: query
          description: Filter by completion status
          schema:
            type: boolean
        - name: tag
          in: query
          description: Filter by tag
          schema:
            type: string
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high]
        - name: isFavorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
        - name: isArchived
          in: query
          description: Filter by archive status
          schema:
            type: boolean
        - name: isDeleted
          in: query
          description: Filter by deletion status
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Todos
      summary: Create a new todo
      description: Create a new todo with title, description, and metadata
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /todos/{id}:
    get:
      tags:
        - Todos
      summary: Get todo by ID
      description: Retrieve a specific todo by its ID
      operationId: getTodoById
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Todos
      summary: Update todo
      description: Update an existing todo
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Todos
      summary: Delete todo (move to trash)
      description: Soft delete a todo by moving it to trash
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
        - name: permanent
          in: query
          description: Permanently delete without moving to trash
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /todos/{id}/archive:
    post:
      tags:
        - Todos
      summary: Archive todo
      description: Move todo to archive
      operationId: archiveTodo
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      responses:
        '200':
          description: Todo archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /todos/{id}/restore:
    post:
      tags:
        - Todos
      summary: Restore todo
      description: Restore todo from trash or archive
      operationId: restoreTodo
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      responses:
        '200':
          description: Todo restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /templates:
    get:
      tags:
        - Templates
      summary: Get all templates
      description: Retrieve a list of all note templates
      operationId: getTemplates
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteTemplate'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Templates
      summary: Create a new template
      description: Create a new note template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteTemplateInput'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteTemplate'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /templates/{id}:
    get:
      tags:
        - Templates
      summary: Get template by ID
      description: Retrieve a specific template by its ID
      operationId: getTemplateById
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteTemplate'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Templates
      summary: Delete template
      description: Delete a template by ID
      operationId: deleteTemplate
      parameters:
        - name: id
          in: path
          required: true
          description: Template ID
          schema:
            type: string
      responses:
        '204':
          description: Template deleted successfully
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Attachment:
      type: object
      required:
        - id
        - name
        - size
        - type
        - url
        - uploadedAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "att_1234567890"
        name:
          type: string
          description: File name
          example: "document.pdf"
        size:
          type: integer
          description: File size in bytes
          example: 1048576
        type:
          type: string
          description: MIME type
          example: "application/pdf"
        url:
          type: string
          description: File URL or path
          example: "/uploads/document.pdf"
        uploadedAt:
          type: string
          format: date-time
          description: Upload timestamp
          example: "2026-04-13T10:00:00Z"
...
    Todo:
      type: object
      required:
        - id
        - title
        - description
        - completed
        - priority
        - isFavorite
        - isArchived
        - isDeleted
        - attachments
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "9876543210"
        title:
          type: string
          description: Todo title
          example: "Complete project documentation"
        description:
          type: string
          description: Todo description (supports Markdown)
          example: "Write comprehensive documentation for the API endpoints"
        completed:
          type: boolean
          description: Completion status
          example: false
        tags:
          type: array
          items:
            type: string
          description: List of tags
          example: ["documentation", "urgent"]
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: "high"
        isFavorite:
          type: boolean
          description: Favorite status
          example: true
        isArchived:
          type: boolean
          description: Archive status
          example: false
        isDeleted:
          type: boolean
          description: Deletion status (in trash)
          example: false
        deletedAt:
          type: string
          format: date-time
          description: Deletion timestamp
          example: "2026-04-13T12:00:00Z"
        dueDate:
          type: string
          format: date-time
          description: Due date
          example: "2026-04-20T17:00:00Z"
        reminder:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2026-04-19T09:00:00Z"
        location:
          type: object
          properties:
            lat:
              type: number
              format: double
              example: 37.7749
            lng:
              type: number
              format: double
              example: -122.4194
            address:
              type: string
              example: "Office Building A, Room 301"
        schedule:
          type: object
          properties:
            repeat:
              type: string
              enum: [none, daily, weekly, monthly]
              example: "weekly"
            endDate:
              type: string
              format: date-time
              example: "2026-12-31T23:59:59Z"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: List of file attachments
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-04-13T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-04-13T09:15:00Z"
    TodoInput:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          description: Todo title
          example: "Complete project documentation"
        description:
          type: string
          description: Todo description (supports Markdown)
          example: "Write comprehensive documentation for the API endpoints"
        completed:
          type: boolean
          description: Completion status
          example: false
          default: false
        tags:
          type: array
          items:
            type: string
          description: List of tags
          example: ["documentation", "urgent"]
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: "medium"
          default: "medium"
        isFavorite:
          type: boolean
          description: Favorite status
          example: false
          default: false
        dueDate:
          type: string
          format: date-time
          description: Due date
          example: "2026-04-20T17:00:00Z"
        reminder:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2026-04-19T09:00:00Z"
        location:
          type: object
          properties:
            lat:
              type: number
              format: double
            lng:
              type: number
              format: double
            address:
              type: string
        schedule:
          type: object
          properties:
            repeat:
              type: string
              enum: [none, daily, weekly, monthly]
              default: "none"
            endDate:
              type: string
              format: date-time
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: List of file attachments
    NoteTemplate:
      type: object
      required:
        - id
        - name
        - content
        - tags
        - category
        - createdAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "tmpl_123"
        name:
          type: string
          description: Template name
          example: "Meeting Notes"
        content:
          type: string
          description: Template content (Markdown)
          example: "# Meeting Notes\n\n**Date:** \n**Attendees:** \n\n## Agenda\n- \n\n## Discussion\n\n## Action Items\n- [ ] "
        tags:
          type: array
          items:
            type: string
          description: Default tags
          example: ["meeting", "work"]
        category:
          type: string
          description: Template category
          example: "Work"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-04-13T08:30:00Z"
    NoteTemplateInput:
      type: object
      required:
        - name
        - content
        - category
      properties:
        name:
          type: string
          description: Template name
          example: "Meeting Notes"
        content:
          type: string
          description: Template content (Markdown)
          example: "# Meeting Notes\n\n**Date:** \n**Attendees:** "
        tags:
          type: array
          items:
            type: string
          description: Default tags
          example: ["meeting"]
        category:
          type: string
          description: Template category
          example: "Work"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Title is required"
        details:
          type: object
          description: Additional error details
          additionalProperties: true